+++ pkgs/native_assets_builder/test/build_runner/build_runner_asset_id_test.dart
@@ -22,7 +22,7 @@ void main() async {
 
       {
         final result = await build(packageUri, logger, dartExecutable);
-        expect(result.errors, isNotEmpty);
+        expect(result.success, false);
       }
     });
   });