+++ pkgs/native_assets_builder/test/build_runner/build_runner_build_output_format_test.dart
@@ -26,7 +26,7 @@ void main() async {
 
         {
           final result = await build(packageUri, logger, dartExecutable);
-          expect(result.errors, isNotEmpty);
+          expect(result.success, false);
         }
       });
     });