+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -3,8 +3,9 @@
 - **Breaking change** `NativeAssetsBuildRunner`s methods now return an object
   ([#105](https://github.com/dart-lang/native/issues/105)).
 - **Breaking change** `NativeAssetsBuildRunner`s methods now return value now
-  contains a list of errors instead of it throwing
-  ([#106](https://github.com/dart-lang/native/issues/106)).
+  contain a success bool instead of throwing
+  ([#106](https://github.com/dart-lang/native/issues/106)). Error messages are
+  streamed to the logger.
 - Use an `out/` sub directory for building native assets
   ([#98](https://github.com/dart-lang/native/issues/98)).
 - Check asset ids on having having a package uri with the owning package